Most issues fixed
Now the main program works fine, there's just a slight issue after we close the viewer - the next time you open a shape, it will be blank, but after that it'll work fine. Modified all programs to display shapes instantaneously (except Selin's program, but I have made it faster). There is now a .bat file to run the program in the root folder.